+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ x.x.x (TBD)
 * Added transformations for the Panel class (https://grafana.com/docs/grafana/next/panels/transformations/types-options/#transformation-types-and-options)
 * Added Worldmap panel (https://grafana.com/grafana/plugins/grafana-worldmap-panel/)
 * Added missing fill gradient to Graph panel
+* Added missing align to graph panel
 
 Changes
 -------